[Feat ✏️] MediaPlayer Click Event 구현
