+++ src/components/PostCard/components/MediaPlayer/MediaPlayer.Types.ts
@@ -1,0 +1,3 @@
+export interface MediaPlayerProps {
+  isBlock?: boolean
+}
